updating requirements and adding new folder at \src
