--- src/labeling_llm/requirements.txt
@@ -1,5 +1,0 @@
-pandas==2.2.2
-torch==2.2.2
-transformers==4.41.0
-transformers_stream_generator==0.0.5
-tiktoken==0.9.0
